Fixed an issue in escapeXML() whereby if the supplied value is undefined, a JavaScript error was thrown.
I have implemented the fix in a generic fashion, i.e. it will now behave correctly for any non-string value that doesn't implement a toString() method (not just undefined values).  If toString() is available, then this will be used to convert the value to a string prior to performing the necessary replacements, otherwise we catch the resulting JavaScript error and return an empty string.

This works around an IE8 issue, whereby node.node.className.baseVal is not defined, resulting in a 'class' attribute set to undefined.  I have logged the underlying issue as issue #53, but this fix means that we should no longer generate a JavaScript error when this situation occurs.